DtD

Documentation for DtD.

DtD is calculated as the difference between the market value of the assets of the firm and the face value of its debt, scaled by the standard deviation of the firm's asset value. While the face value of the debt of the firm is known, the market value of the assets is not.

Exploiting the option nature of equity as a European call option on the underlying assets of a firm, the Merton Model (1974) derives the implied market value of the firm's assets and its volatility by solving the Black-Scholes (BS) equation backwards.

DtD.dtd

DtD.dtd — Type

This function implements the Merton Model (1974) to compute a measure of credit risk of a firm: Distance to default (DtD). DtD indicates how many standard deviations is a firm away from the default point.

Inputs: 

mcap: Is a scalar specifying the market capital of the firm.               

debt: Is a scalar specifying the threshold level of debt for the firm below which the firm will default. Should be a non-zero number.

 vol: Is a scalar specifying the equity volatility of the firm.

   r: Is a scalar specifying the annualized risk free interest rate.

Outputs:

      a dtd object is returned, which has 3 elements:
    
    dtd.v: Distance to default value of the firm                                
   
    asset.v: Estimated asset value of the firm                                    
   
    sigma.v: Estimated volatility of the asset value of the firm

Example:

julia> dtd(100, 70, 0.3, 0.1) |> print
dtd_v: 3.33333306534595
asset_v: 3.33333306534595
sigma_v: 3.33333306534595